--- spring-boot-3-demo-app/src/test/resources/application-test-retry.yml
@@ -1,11 +1,0 @@
-app:
-    external-base-url: "http://localhost:${wiremock.server.port}/"
-    retries: 1
-
-logging:
-    appender:
-        name: CONSOLE
-    level:
-        org.testcontainers: INFO # In order to troubleshoot issues with Testcontainers, increase the logging level to DEBUG
-        com.github.dockerjava: WARN
-        com.github.dockerjava.zerodep.shaded.org.apache.hc.client5.http.wire: OFF
